debug

+++ hc_visc_thb.c
@@ -478,7 +478,7 @@ void thb_postprocess_read(struct thb_solution *sol, struct hc_parameters *p){
     }
   }
   /* sort based on residuals */
-  int *indices = (int *) malloc(iline*sizeof(int));
+  size_t *indices = (size_t *) malloc(iline*sizeof(size_t));
   for(int i=0;i<iline;i++) indices[i] = i;
   gsl_sort_index(indices, residuals, 1, iline);
   int median_index = indices[iline/2];
